solver almost done :D
so close to finishing the solver. need to make a for loop in the solver function which partitions both the boundary condition k matrix and the original one. today i learned the former is used when finding the displacements, and the latter is used to find the forces, so i need to implement that. when thats done, i believe the solver will finally work